fix: 리뷰 리스트 속 아이템의 별점이 평점이 아닌 그 리뷰의 별점을 반환하도록 수정

+++ src/main/java/ssu/eatssu/domain/review/dto/MealReviewResponse.java
@@ -1,17 +1,16 @@
 package ssu.eatssu.domain.review.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import ssu.eatssu.domain.review.entity.Review;
-import ssu.eatssu.domain.review.entity.ReviewMenuLike;
-
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import ssu.eatssu.domain.review.entity.Review;
+import ssu.eatssu.domain.review.entity.ReviewMenuLike;
 
 @AllArgsConstructor
 @Builder
@@ -31,7 +30,7 @@ public class MealReviewResponse {
     private String writerNickname;
 
     @Schema(description = "평점", example = "4")
-    private Double rating;
+    private Integer rating;
 
     @Schema(description = "리뷰 작성 날짜(format = yyyy-MM-dd)", example = "2023-04-07")
     private LocalDate writtenAt;
@@ -60,7 +59,7 @@ public class MealReviewResponse {
 
     public static MealReviewResponse from(Review review,
                                           Long userId,
-                                          List<ValidMenuForViewResponse.MenuDto> validMenus,Double rating) {
+        List<ValidMenuForViewResponse.MenuDto> validMenus, Integer rating) {
         List<String> imageUrls = new ArrayList<>();
         review.getReviewImages().forEach(i -> imageUrls.add(i.getImageUrl()));
